+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Change Log
 * Refatora o sistema para lidar com o novo código único do pneu e com o código do cliente
 * Bloqueia resources ligados a pneu para os apps com version code menor ou igual a 51
 * Seta `ResetAbandonedTimer` para resetar timer de abandono de uma connection
+* Altera insert do pneu para retornar um `AbstractResponse`
 
 ### Bug Fixes
 * Aprimora fluxo de inserção de um modelo de checklist